refactor(core): split crud/analysis into per-entity modules
crud/analysis.py had grown to ~1168 lines mixing CRUD for four unrelated
entities behind one flat module: dead-code findings, code-health
(findings/metrics/snapshots), refactoring suggestions, and coverage
files.

Split it into a crud/analysis/ package grouped by entity:

- dead_code.py    dead-code finding persistence
- health.py       health findings/metrics/snapshots/governance
- refactoring.py  refactoring-suggestion persistence
- coverage.py     coverage-file persistence

The one cross-entity helper (_finding_file_path, used by both dead_code
and refactoring) moves to the existing crud/_shared.py. Pure structural
move: every function relocates verbatim and the three lazy imports in the
health functions stay function-local.

__init__.py re-exports the public functions with an explicit __all__, so
the crud facade (from .analysis import *) and the direct test importers
keep the same public surface. A grep confirmed nothing relied on the ORM
model names that the previous no-__all__ star-export incidentally leaked
through the crud namespace.
